Create groovis.groovy
this is a goofy little proof of concept of saving a object to a redis set using jedis and a json string; reading said string back and seeing it again as an object with an iterator

its about as simple a redis proof of concept you can do but its a basis of a redis stack

you really should be using redisson instead of this. Its a great library and through this great community its only getting better.